projects/VaultCraft/apps/backend/app/price_provider.py
from __future__ import annotations

from typing import Dict, List
import time

from .hyper_client import HyperHTTP
from .settings import Settings, settings
from .cache import TTLCache


class PriceProvider:
    def get_index_prices(self, symbols: List[str]) -> Dict[str, float]:
        raise NotImplementedError


class RestPriceProvider(PriceProvider):
    def __init__(self, api_base: str | None = None, timeout: float | None = None):
        self.http = HyperHTTP(api_base=api_base or settings.HYPER_API_URL, rpc_url=settings.HYPER_RPC_URL, timeout=timeout or settings.PRICE_TIMEOUT)

    def get_index_prices(self, symbols: List[str]) -> Dict[str, float]:
        return self.http.get_index_prices(symbols)


class SDKPriceProvider(PriceProvider):
    def __init__(self, api_base: str | None = None, timeout: float | None = None):
        self.api_base = api_base or settings.HYPER_API_URL
        self.timeout = timeout or settings.PRICE_TIMEOUT
        self._info = None

    @staticmethod
    def available() -> bool:
        try:
            import hyperliquid  # noqa: F401
            return True
        except Exception:
            return False

    def _get_info(self):
        if self._info is None:
            from hyperliquid.info import Info  # type: ignore
            # Avoid opening WS in SDK to keep it lightweight in API calls
            self._info = Info(base_url=self.api_base, skip_ws=True, timeout=self.timeout)
        return self._info

    def get_index_prices(self, symbols: List[str]) -> Dict[str, float]:
        info = self._get_info()
        mids = info.all_mids()
        # SDK returns list of {"name": sym, "mid": price} or similar; accept common shapes
        out: Dict[str, float] = {}
        if isinstance(mids, list):
            for item in mids:
                name = item.get("name") or item.get("symbol")
                mid = item.get("mid") or item.get("price")
                if name in symbols and mid is not None:
                    out[name] = float(mid)
        elif isinstance(mids, dict):
            # sometimes as {sym: price}
            for k, v in mids.items():
                if k in symbols:
                    out[k] = float(v)
        return out


class PriceRouter(PriceProvider):
    def __init__(self):
        # Rebuild settings to reflect current env at instantiation (useful in tests)
        env = Settings()
        self.sdk_enabled = bool(env.ENABLE_HYPER_SDK)
        self.sdk = SDKPriceProvider()
        self.rest = RestPriceProvider()

    def get_index_prices(self, symbols: List[str]) -> Dict[str, float]:
        symbols = [s for s in symbols if s]
        if not symbols:
            return {}
        if self.sdk_enabled and self.sdk.available():
            try:
                # If HyperHTTP.get_index_prices has been monkeypatched (tests),
                # honor the patch by preferring REST path to keep determinism.
                try:
                    from .hyper_client import HyperHTTP as _H  # local import to avoid cycles
                    patched = getattr(_H.get_index_prices, "__module__", "") != "app.hyper_client"
                except Exception:
                    patched = False
                if not patched:
                    data = self.sdk.get_index_prices(symbols)
                else:
                    raise RuntimeError("prefer_rest_due_to_patched_hyperhttp")
                if data:
                    return data
            except Exception:
                pass
        # fallback to REST; outer layer may fallback to deterministic
        return self.rest.get_index_prices(symbols)


class CachedPriceRouter(PriceProvider):
    def __init__(self, router: PriceRouter | None = None, ttl_seconds: float | None = None):
        self.router = router or PriceRouter()
        ttl = ttl_seconds if ttl_seconds is not None else float(getattr(settings, "PRICE_CACHE_TTL", 5.0))
        self.cache = TTLCache[str, Dict[str, float]](ttl_seconds=ttl)
        self.last_good: Dict[str, Dict[str, float]] = {}

    def get_index_prices(self, symbols: List[str]) -> Dict[str, float]:
        key = ",".join(sorted([s for s in symbols if s]))
        cached = self.cache.get(key)
        if cached is not None:
            return cached
        attempts = int(getattr(settings, "PRICE_RETRIES", 1)) + 1
        backoff = float(getattr(settings, "PRICE_RETRY_BACKOFF_SEC", 0.2))
        last_err: Exception | None = None
        data: Dict[str, float] | None = None
        for i in range(attempts):
            try:
                data = self.router.get_index_prices(symbols)
                break
            except Exception as e:
                last_err = e
                if i < attempts - 1:
                    try:
                        time.sleep(backoff * (2 ** i))
                    except Exception:
                        pass
        if data is None or not data:
            # return last_good if available
            lg = self.last_good.get(key)
            if lg:
                return lg
            # propagate last error for outer handler
            if last_err:
                raise last_err
            return {}
        if data:
            self.cache.set(key, data)
            self.last_good[key] = data
        return data
