projects/VaultCraft/apps/backend/pyproject.toml
[project]
name = "vaultcraft-backend"
version = "0.0.1"
description = "VaultCraft v0 backend (metrics, indexing, APIs)"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic-settings",
  "httpx",
  "web3",
  "sqlalchemy",
  "aiosqlite",
  "hyperliquid-python-sdk>=0.19.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

